backend/requirements.txt
torch==2.4.0
torchvision==0.19.0
pillow==10.4.0
fastapi==0.115.0
uvicorn[standard]==0.30.5
python-multipart==0.0.9
httpx==0.27.2
python-jose[cryptography]==3.3.0
PyJWT[crypto]==2.9.0  
tqdm==4.66.4
flask
flask-cors
requests